requirements.txt
pyautogui==0.9.53
httpx==0.22.0
python-dotenv==0.20.0